+++ src/has-paid.ts
@@ -1,17 +1,13 @@
 import { useState, useEffect, useRef } from 'react'
 
-import { MonetizationProgressEvent, MonetizationPendingEvent } from './types'
+import { useListener } from './listener'
 
 export function useHasPaid(): boolean {
   const [hasPaid, setHasPaid] = useState(false)
   const requestIdRef = useRef<string>()
 
-  useEffect(() => {
-    const { monetization } = document
-
-    if (!monetization) return
-
-    function handlePending(event: MonetizationPendingEvent): void {
+  useListener({
+    onPending(event) {
       const { requestId } = event.detail
 
       if (requestIdRef.current !== requestId) {
@@ -19,23 +15,12 @@ export function useHasPaid(): boolean {
       }
 
       requestIdRef.current = requestId
-    }
-
-    function handleProgress(event: MonetizationProgressEvent): void {
+    },
+    onProgress(event) {
       requestIdRef.current = event.detail.requestId
       setHasPaid(true)
     }
-
-    // monetizationstop events can include requestId for new paymentPointer
-
-    monetization.addEventListener('monetizationpending', handlePending)
-    monetization.addEventListener('monetizationprogress', handleProgress)
-
-    return (): void => {
-      monetization.removeEventListener('monetizationpending', handlePending)
-      monetization.removeEventListener('monetizationprogress', handleProgress)
-    }
-  }, [])
+  })
 
   return hasPaid
 }