+++ src/index.ts
@@ -1,3 +1,4 @@
+export * from './listener'
 export * from './value'
 export * from './counter'
 export * from './status'