+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.1.3] - 2019-11-2
+### Added
+- `useListener`.
+
+### Changed
+- Using [dacioromero/wm-types] over local types.
+- Using `useListener` in all hooks.
+
+
 ## [0.1.2] - 2019-11-24
 ### Changed
 - Reset `useHasPaid` on `requestId` change.
@@ -19,6 +28,7 @@
 - `useHasPaid`
 
 
+[0.1.3]: https://github.com/dacioromero/react-hook-wm/compare/0.1.2...0.1.3
 [0.1.2]: https://github.com/dacioromero/react-hook-wm/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/dacioromero/react-hook-wm/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/dacioromero/react-hook-wm/releases/tag/0.1.0